+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- `git pkgs history` now works without a package argument to show all dependency changes
+
 ## [0.1.0] - 2026-01-01
 
 - Initial release